Make cmdline programs pip installable
