+++ setup.py
@@ -21,4 +21,11 @@ setup(
         "xarray",
         "dask",
     ],
+    entry_points={
+        "console_scripts": [
+            "fileio = unseen.fileio._main",
+            "independence_test = unseen.independence_test._main",
+            "similarity_test = unseen.similarity_test._main",
+            "bias_correction = unseen.bias_correction._main"]}
+            
 )
