+++ unseen/similarity_test.py
@@ -1,12 +1,4 @@
-"""Command line program for similarity testing."""
-
-import sys
-
-script_dir = sys.path[0]
-repo_dir = "/".join(script_dir.split("/")[:-1])
-module_dir = repo_dir + "/unseen"
-sys.path.insert(1, module_dir)
-sys.path.insert(1, "/home/599/dbi599/xks/")
+"""Funcitons and command line program for similarity testing."""
 
 import argparse
 
@@ -49,8 +41,47 @@ def univariate_ks_test(fcst_stacked, obs_stacked, var):
     return ds
 
 
-def _main(args):
+def _parse_command_line():
+    """Parse the command line for input agruments"""
+
+    parser = argparse.ArgumentParser(
+        description=__doc__, formatter_class=argparse.RawDescriptionHelpFormatter
+    )
+
+    parser.add_argument("fcst_file", type=str, help="Forecast file")
+    parser.add_argument("obs_file", type=str, help="Observations file")
+    parser.add_argument("var", type=str, help="Variable name")
+    parser.add_argument("outfile", type=str, help="Output file")
+
+    parser.add_argument(
+        "--dask_config", type=str, help="YAML file specifying dask client configuration"
+    )
+
+    parser.add_argument(
+        "--reference_time_period",
+        type=str,
+        nargs=2,
+        default=None,
+        help="Start and end date (YYYY-MM-DD format)",
+    )
+    parser.add_argument(
+        "--output_chunks",
+        type=str,
+        nargs="*",
+        action=general_utils.store_dict,
+        default={},
+        help="Chunks for writing data to file (e.g. init_date=-1 lead_time=-1)",
+    )
+
+    args = parser.parse_args()
+    
+    return args
+
+
+def _main():
     """Run the command line program."""
+
+    args = _parse_command_line()
 
     if args.dask_config:
         client = dask_setup.launch_client(args.dask_config)
@@ -85,37 +116,5 @@ def _main(args):
 
 
 if __name__ == "__main__":
-    parser = argparse.ArgumentParser(
-        description=__doc__, formatter_class=argparse.RawDescriptionHelpFormatter
-    )
+    _main()
 
-    parser.add_argument("fcst_file", type=str, help="Forecast file")
-    parser.add_argument("obs_file", type=str, help="Observations file")
-    parser.add_argument("var", type=str, help="Variable name")
-    #    parser.add_argument("test", type=str,
-    #                        choices=('kolmogorov_smirnov'),
-    #                        help="Similarity test")
-    parser.add_argument("outfile", type=str, help="Output file")
-
-    parser.add_argument(
-        "--dask_config", type=str, help="YAML file specifying dask client configuration"
-    )
-
-    parser.add_argument(
-        "--reference_time_period",
-        type=str,
-        nargs=2,
-        default=None,
-        help="Start and end date (YYYY-MM-DD format)",
-    )
-    parser.add_argument(
-        "--output_chunks",
-        type=str,
-        nargs="*",
-        action=general_utils.store_dict,
-        default={},
-        help="Chunks for writing data to file (e.g. init_date=-1 lead_time=-1)",
-    )
-
-    args = parser.parse_args()
-    _main(args)
